+++ packages/perseus/src/widgets/interactive-graphs/locked-figures/locked-vector.tsx
@@ -3,10 +3,12 @@ import * as React from "react";
 
 import {Vector} from "../graphs/components/vector";
 
+import {strokeWeights} from "./utils";
+
 import type {LockedVectorType} from "@khanacademy/perseus-core";
 
 const LockedVector = (props: LockedVectorType) => {
-    const {color, points, ariaLabel} = props;
+    const {color, points, weight, ariaLabel} = props;
     const [tail, tip] = points;
 
     const hasAria = !!ariaLabel;
@@ -18,7 +20,12 @@ const LockedVector = (props: LockedVectorType) => {
             aria-hidden={!hasAria}
             role="img"
         >
-            <Vector tail={tail} tip={tip} color={lockedFigureColors[color]} />
+            <Vector
+                tail={tail}
+                tip={tip}
+                color={lockedFigureColors[color]}
+                strokeWidth={strokeWeights[weight]}
+            />
         </g>
     );
 };